bin: integrate fail to tikv ctl (#3364)
